# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "openai-compatible-proxy",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/openai-compatible-proxy",
    "paths": {
        "/api/show": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ollama"
                ],
                "summary": "Get details of a specific model by name",
                "parameters": [
                    {
                        "description": "Model name",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OllamaShowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OllamaModel"
                        }
                    }
                }
            }
        },
        "/api/tags": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ollama"
                ],
                "summary": "Get tags from models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OllamaResponse"
                        }
                    }
                }
            }
        },
        "/api/version": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ollama"
                ],
                "summary": "Get version information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OllamaVersionResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Get health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/model": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "Get models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Model"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/model/lm-studio": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "Fetch and save models from LM Studio",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Model"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/model/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "summary": "Get model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Model"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Model": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string"
                },
                "compatibilityType": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "maxContextLength": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "publisher": {
                    "type": "string"
                },
                "quantization": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "request.OllamaShowRequest": {
            "type": "object",
            "required": [
                "model"
            ],
            "properties": {
                "model": {
                    "type": "string"
                }
            }
        },
        "response.Details": {
            "type": "object",
            "properties": {
                "families": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "family": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "parameter_size": {
                    "type": "string"
                },
                "parent_model": {
                    "type": "string"
                },
                "quantization_level": {
                    "type": "string"
                }
            }
        },
        "response.OllamaModel": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "details": {
                    "$ref": "#/definitions/response.Details"
                },
                "digest": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "model_info": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "modified_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "response.OllamaResponse": {
            "type": "object",
            "properties": {
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.OllamaModel"
                    }
                }
            }
        },
        "response.OllamaVersionResponse": {
            "type": "object",
            "properties": {
                "version": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "Bearer": {
            "description": "Generated by /authentication",
            "type": "apiKey",
            "name": "X-AUTHORIZATION",
            "in": "header"
        }
    }
}